Fixes issues on Cronjob
Fixes an issue when requesting the
storefront.graduated_prices_service-service on cronjob.

Due to the createEvents() method, this service is being extended
by the plugin. However, it causes an ugly crash because its
dependencies, e.g. service swaguserprice.userprice are not loaded
or registered at cronjob. This ensures these services are
available and that the decoration of the storefront-services
scuceeds.